requirements.txt
asgiref==3.4.1
Django==3.2.6
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.28
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
